# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Relation API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v0/products": {
            "get": {
                "tags": [
                    "Products Data"
                ],
                "summary": "Get Products",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Products"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/products?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post Products Data"
                ],
                "summary": "Post Products",
                "operationId": "postProducts",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Jack Daniel"
                                    },
                                    "license_name": {
                                        "type": "string",
                                        "example": "HACCP"
                                    },
                                    "tag_name": {
                                        "type": "string",
                                        "example": "Pilsner"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product has been created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Products"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v0/product-tags": {
            "get": {
                "tags": [
                    "Product's Tag Data"
                ],
                "summary": "Get Product's and Tags",
                "operationId": "getProductTags",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "example": "Sony"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Products"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/product-tags?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v0/product-licenses": {
            "get": {
                "tags": [
                    "Product's License Data"
                ],
                "summary": "Get Product's and Licenses",
                "operationId": "getProductLicenses",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    },
                    {
                        "name": "license",
                        "in": "query",
                        "example": "iBM"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Products"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/product-licenses?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v0/products/{id}": {
            "get": {
                "tags": [
                    "Products by Id Data"
                ],
                "summary": "Get Product by Id",
                "operationId": "getProductId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Products"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/products?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Update Product Data"
                ],
                "summary": "Update Product Data",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Vivo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product has been updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Products"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Product Data"
                ],
                "summary": "Delete Product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product has been deleted"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Products": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Smart Window"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-05-13T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-05-13T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    }
}